GEOMESA-1832 Refactor WPS (#1523)
* Modules and install files for accumulo and hbase processes
* Added license:check to validate lifecycle
