+++ geomesa-accumulo/geomesa-accumulo-datastore/pom.xml
@@ -134,6 +134,11 @@
 
         <!-- test dependencies -->
         <dependency>
+            <groupId>org.locationtech.geomesa</groupId>
+            <artifactId>geomesa-process_${scala.binary.version}</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-log4j12</artifactId>
         </dependency>
