+++ geomesa-accumulo/geomesa-accumulo-datastore/src/main/scala/org/locationtech/geomesa/accumulo/iterators/BinAggregatingIterator.scala
@@ -26,6 +26,7 @@ import org.locationtech.geomesa.features.SerializationOption.SerializationOption
 import org.locationtech.geomesa.features.SerializationType.SerializationType
 import org.locationtech.geomesa.features.kryo.KryoBufferSimpleFeature
 import org.locationtech.geomesa.features.{ScalaSimpleFeature, SimpleFeatureDeserializers}
+import org.locationtech.geomesa.filter.function.BinaryOutputEncoder.BIN_ATTRIBUTE_INDEX
 import org.locationtech.geomesa.filter.function.{BasicValues, BinaryOutputEncoder, Convert2ViewerFunction, ExtendedValues}
 import org.locationtech.geomesa.utils.geotools.RichSimpleFeatureType.RichSimpleFeatureType
 import org.locationtech.geomesa.utils.text.WKTUtils
@@ -356,7 +357,6 @@ object BinAggregatingIterator extends LazyLogging {
 
   import org.locationtech.geomesa.index.conf.QueryHints.RichHints
 
-  val BIN_ATTRIBUTE_INDEX = 0 // index of 'bin' attribute in BIN_SFT
   private val zeroPoint = WKTUtils.read("POINT(0 0)")
 
   val DEFAULT_PRIORITY = 25
