+++ geomesa-accumulo/geomesa-accumulo-datastore/src/test/scala/org/locationtech/geomesa/accumulo/process/knn/NearestNeighborsPQTest.scala
@@ -8,7 +8,6 @@
 
 package org.locationtech.geomesa.accumulo.process.knn
 
-
 import org.geotools.factory.Hints
 import org.geotools.feature.DefaultFeatureCollection
 import org.geotools.feature.simple.SimpleFeatureBuilder
