+++ geomesa-accumulo/geomesa-accumulo-dist/pom.xml
@@ -54,6 +54,13 @@
         </dependency>
         <dependency>
             <groupId>org.locationtech.geomesa</groupId>
+            <artifactId>geomesa-accumulo-process_${scala.binary.version}</artifactId>
+            <classifier>install</classifier>
+            <version>${project.version}</version>
+            <type>tar.gz</type>
+        </dependency>
+        <dependency>
+            <groupId>org.locationtech.geomesa</groupId>
             <artifactId>geomesa-blobstore-gs-plugin_${scala.binary.version}</artifactId>
             <classifier>install</classifier>
             <version>${project.version}</version>
@@ -72,13 +79,6 @@
             <classifier>install</classifier>
             <version>${project.version}</version>
             <type>tar.gz</type>
-        </dependency>
-
-        <!-- GeoMesa Process jar standalone since it requires WPS and we dont -->
-        <!-- want to force people to install WPS -->
-        <dependency>
-            <groupId>org.locationtech.geomesa</groupId>
-            <artifactId>geomesa-process_${scala.binary.version}</artifactId>
         </dependency>
 
         <!-- Web Services (dist) -->
