+++ geomesa-accumulo/geomesa-accumulo-dist/src/main/assemblies/component.xml
@@ -41,10 +41,10 @@
             <useTransitiveDependencies>false</useTransitiveDependencies>
             <includes>
                 <include>org.locationtech.geomesa:geomesa-accumulo-gs-plugin*</include>
+                <include>org.locationtech.geomesa:geomesa-accumulo-process*</include>
                 <include>org.locationtech.geomesa:geomesa-blobstore-gs-plugin*</include>
                 <include>org.locationtech.geomesa:geomesa-stream-gs-plugin*</include>
                 <include>org.locationtech.geomesa:geomesa-geojson-gs-plugin*</include>
-                <include>org.locationtech.geomesa:geomesa-process*</include>
             </includes>
         </dependencySet>
         <dependencySet>
