+++ geomesa-accumulo/geomesa-accumulo-process/src/main/resources/META-INF/services/org.geotools.process.ProcessFactory
@@ -1,0 +1,1 @@
+org.locationtech.geomesa.accumulo.process.ProcessFactory
