+++ geomesa-accumulo/pom.xml
@@ -19,6 +19,7 @@
         <module>geomesa-accumulo-gs-plugin</module>
         <module>geomesa-accumulo-jobs</module>
         <module>geomesa-accumulo-native-api</module>
+        <module>geomesa-accumulo-process</module>
         <module>geomesa-accumulo-raster</module>
         <module>geomesa-accumulo-security</module>
         <module>geomesa-accumulo-spark</module>
