+++ geomesa-hbase/geomesa-hbase-dist/pom.xml
@@ -53,6 +53,13 @@
                 </exclusion>
             </exclusions>
         </dependency>
+        <dependency>
+            <groupId>org.locationtech.geomesa</groupId>
+            <artifactId>geomesa-hbase-process_${scala.binary.version}</artifactId>
+            <classifier>install</classifier>
+            <version>${project.version}</version>
+            <type>tar.gz</type>
+        </dependency>
     
         <!-- tools -->
         <dependency>
