+++ geomesa-hbase/geomesa-hbase-process/src/main/assemblies/assembly.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<assembly xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2"
+          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+          xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2 http://maven.apache.org/xsd/assembly-1.1.2.xsd">
+  <id>install</id>
+  <formats>
+    <format>tar.gz</format>
+  </formats>
+  <includeBaseDirectory>false</includeBaseDirectory>
+  <baseDirectory>geomesa-hbase_${scala.binary.version}-${project.version}</baseDirectory>
+
+  <componentDescriptors>
+    <componentDescriptor>src/main/assemblies/component.xml</componentDescriptor>
+  </componentDescriptors>
+
+  <dependencySets>
+    <dependencySet>
+      <outputDirectory>/</outputDirectory>
+      <!--Set to 'true' to include geomesa-hbase-process jar-->
+      <useProjectArtifact>false</useProjectArtifact>
+      <scope></scope>
+    </dependencySet>
+  </dependencySets>
+
+</assembly>
