+++ geomesa-hbase/pom.xml
@@ -42,6 +42,7 @@
         <module>geomesa-hbase-distributed-runtime</module>
         <module>geomesa-hbase-gs-plugin</module>
         <module>geomesa-hbase-native-api</module>
+        <module>geomesa-hbase-process</module>
         <module>geomesa-hbase-tools</module>
         <module>geomesa-hbase-spark</module>
         <module>geomesa-hbase-spark-runtime</module>
